+++ classes/EP_Debug_Bar_ElasticPress.php
@@ -25,25 +25,27 @@ class EP_Debug_Bar_ElasticPress extends \Debug_Bar_Panel {
 	public $title;
 
 	/**
+	 * Common panel instance
+	 *
+	 * @var \DebugBarElasticPress\CommonPanel
+	 */
+	protected $common_panel;
+
+	/**
 	 * Initial debug bar stuff
 	 */
 	public function init() {
 		$this->title( esc_html__( 'ElasticPress', 'debug-bar-elasticpress' ) );
 
-		add_action( 'wp_enqueue_scripts', array( $this, 'enqueue_scripts_styles' ) );
-		add_action( 'admin_enqueue_scripts', array( $this, 'enqueue_scripts_styles' ) );
+		$this->common_panel = new \DebugBarElasticPress\CommonPanel();
+		$this->common_panel->enqueue_scripts_styles();
 	}
 
 	/**
 	 * Enqueue scripts for front end and admin
 	 */
 	public function enqueue_scripts_styles() {
-		if ( ! is_user_logged_in() ) {
-			return;
-		}
-
-		wp_enqueue_script( 'debug-bar-elasticpress', EP_DEBUG_URL . 'assets/js/main.js', array( 'wp-dom-ready', 'clipboard' ), EP_DEBUG_VERSION, true );
-		wp_enqueue_style( 'debug-bar-elasticpress', EP_DEBUG_URL . 'assets/css/main.css', array(), EP_DEBUG_VERSION );
+		_deprecated_function( __METHOD__, '3.1.0', 'DebugBarElasticPress\EP_Panel::enqueue_scripts_styles()' );
 	}
 
 	/**
@@ -85,15 +87,7 @@ class EP_Debug_Bar_ElasticPress extends \Debug_Bar_Panel {
 			?>
 		</h2>
 
-		<?php if ( function_exists( '\ElasticPress\Utils\is_indexing' ) && \ElasticPress\Utils\is_indexing() ) : ?>
-			<div class="ep-debug-bar-warning">
-				<?php esc_html_e( 'ElasticPress is currently indexing.', 'debug-bar-elasticpress' ); ?>
-			</div>
-		<?php endif; ?>
-
 		<?php
-		$debug_bar_output = new \DebugBarElasticPress\QueryOutput( $queries );
-		$debug_bar_output->render_buttons();
-		$debug_bar_output->render_queries();
+		$this->common_panel->render();
 	}
 }