+++ classes/QueryOutput.php
@@ -52,7 +52,7 @@ class QueryOutput {
 			<a download="debug-bar-elasticpress-report.txt" href="data:text/plain;charset=utf-8,<?php echo rawurlencode( $copy_paste_output ); ?>"  class="button button-primary" id="ep-download-requests-info">
 				<?php esc_html_e( 'Download Requests Info', 'debug-bar-elasticpress' ); ?>
 			</a>
-			<button class="ep-copy-button button qm-button" data-clipboard-text="<?php echo esc_attr( $copy_paste_output ); ?>">
+			<button class="ep-copy-button button bordered-button" data-clipboard-text="<?php echo esc_attr( $copy_paste_output ); ?>">
 				<?php esc_html_e( 'Copy Requests Info to Clipboard', 'debug-bar-elasticpress' ); ?>
 			</button>
 			<span class="ep-copy-button-wrapper__success" style="display: none;">
@@ -66,9 +66,10 @@ class QueryOutput {
 	 * Render the queries
 	 *
 	 * @since 3.0.0
+	 * @param array $args Arguments to adjust the queries list
 	 * @return void
 	 */
-	public function render_queries() {
+	public function render_queries( $args = [] ) {
 		?>
 		<div class="ep-queries-debug-container">
 			<ol class="wpd-queries ep-queries-debug">
@@ -79,7 +80,11 @@ class QueryOutput {
 					<?php
 				} else {
 					foreach ( $this->queries as $query ) {
-						$this->render_query( $query );
+						$type    = $query['args']['ep_query_type'] ?? '';
+						$context = ( ! empty( $args['display_context'] ) && ! empty( $query['args']['ep_context'] ) ) ?
+							$query['args']['ep_context'] :
+							'';
+						$this->render_query( $query, $type, $context );
 					}
 				}
 				?>
@@ -91,11 +96,12 @@ class QueryOutput {
 	/**
 	 * Render a query in a list.
 	 *
-	 * @param array  $query The query info.
-	 * @param string $type The type of the query.
+	 * @param array  $query   The query info.
+	 * @param string $type    The type of the query.
+	 * @param string $context Context of the query (public, admin, ajax, or rest).
 	 * @return void
 	 */
-	public function render_query( $query, $type = '' ) {
+	public function render_query( $query, $type = '', $context = '' ) {
 		$error         = '';
 		$query_time    = ( ! empty( $query['time_start'] ) && ! empty( $query['time_finish'] ) ) ? $query['time_finish'] - $query['time_start'] : false;
 		$result        = wp_remote_retrieve_body( $query['request'] );
@@ -159,6 +165,13 @@ class QueryOutput {
 				<div class="ep-query-type">
 					<strong><?php esc_html_e( 'Type:', 'debug-bar-elasticpress' ); ?></strong>
 					<?php echo esc_html( $type ); ?>
+				</div>
+			<?php endif; ?>
+
+			<?php if ( $context ) : ?>
+				<div class="ep-query-context">
+					<strong><?php esc_html_e( 'Context:', 'debug-bar-elasticpress' ); ?></strong>
+					<?php echo esc_html( $context ); ?>
 				</div>
 			<?php endif; ?>
 
@@ -297,5 +310,99 @@ class QueryOutput {
 		}
 
 		return (string) $value;
+	}
+
+	/**
+	 * Render the additional buttons.
+	 *
+	 * @since 3.1.0
+	 * @return void
+	 */
+	public function render_additional_buttons() {
+		$buttons = [
+			$this->get_explain_query_button(),
+			$this->get_retrieve_raw_document_button(),
+		];
+
+		/**
+		 * Filter the additional buttons.
+		 *
+		 * @since 3.1.0
+		 * @hook ep_debug_bar_additional_buttons
+		 * @param array $buttons Buttons.
+		 * @return array
+		 */
+		apply_filters( 'ep_debug_bar_additional_buttons', $buttons );
+
+		$buttons = array_filter( $buttons );
+
+		if ( empty( $buttons ) ) {
+			return;
+		}
+
+		?>
+		<div class="ep-queries-buttons-wrapper">
+			<?php
+			foreach ( $buttons as $button ) {
+				echo wp_kses_post( $button );
+			}
+			?>
+		</div>
+		<?php
+	}
+
+	/**
+	 * Get the explain query button
+	 *
+	 * @since 3.1.0
+	 * @return string
+	 */
+	public function get_explain_query_button() : string {
+		if ( empty( $this->queries ) ) {
+			return '';
+		}
+
+		$button_link = add_query_arg(
+			[
+				'explain' => '1',
+			],
+			isset( $_SERVER['REQUEST_URI'] ) ? sanitize_text_field( wp_unslash( $_SERVER['REQUEST_URI'] ) ) : ''
+		);
+
+		return sprintf(
+			'<a href="%s" class="ep-explain-query button bordered-button">
+				%s
+			</a>',
+			esc_url( $button_link ),
+			esc_html__( 'Explain queries', 'debug-bar-elasticpress' )
+		);
+	}
+
+	/**
+	 * Get the retrieve raw document button.
+	 *
+	 * @since 3.1.0
+	 * @return string
+	 */
+	public function get_retrieve_raw_document_button() : string {
+		if ( ! is_indexable_singular() ) {
+			return '';
+		}
+
+		$button_link = add_query_arg(
+			[
+				'ep-retrieve-es-document' => '1',
+				'_wpnonce'                => wp_create_nonce( 'ep-retrieve-es-document' ),
+			],
+			isset( $_SERVER['REQUEST_URI'] ) ? sanitize_text_field( wp_unslash( $_SERVER['REQUEST_URI'] ) ) : ''
+		);
+
+		return sprintf(
+			'<a href="%s" class="ep-retrieve-es-document button bordered-button">
+				%s
+			</a>',
+			esc_url( $button_link ),
+			esc_html__( 'Reload and retrieve raw ES document', 'debug-bar-elasticpress' )
+		);
 	}
 }