+++ debug-bar-elasticpress.php
@@ -1,10 +1,10 @@
 <?php
 /**
- * Plugin Name:       Debug Bar ElasticPress
+ * Plugin Name:       ElasticPress Debugging Add-On
  * Plugin URI:        https://wordpress.org/plugins/debug-bar-elasticpress
- * Description:       Extends the debug bar plugin for ElasticPress queries.
+ * Description:       Extends the Query Monitor and Debug Bar plugins for ElasticPress queries.
  * Author:            10up
- * Version:           3.0.0
+ * Version:           3.1.0
  * Author URI:        https://10up.com
  * Requires PHP:      7.0
  * Requires at least: 5.6
@@ -18,7 +18,7 @@
 
 namespace DebugBarElasticPress;
 
-define( 'EP_DEBUG_VERSION', '3.0.0' );
+define( 'EP_DEBUG_VERSION', '3.1.0' );
 define( 'EP_DEBUG_URL', plugin_dir_url( __FILE__ ) );
 define( 'EP_DEBUG_MIN_EP_VERSION', '4.4.0' );
 
@@ -63,9 +63,19 @@ function setup() {
 		return;
 	}
 
-	add_filter( 'debug_bar_panels', $n( 'add_debug_bar_panel' ) );
-	add_filter( 'debug_bar_statuses', $n( 'add_debug_bar_stati' ) );
+	// If Query Monitor is active, do not add the Debug Bar panel (it will be potentially duplicated otherwise)
+	if ( class_exists( '\QM_Collectors' ) ) {
+		\QM_Collectors::add( new QueryMonitorCollector() );
+		add_filter( 'qm/outputter/html', $n( 'register_qm_output' ) );
+		add_action( 'qm/output/enqueued-assets', [ new CommonPanel(), 'enqueue_scripts_styles' ] );
+	} else {
+		add_filter( 'debug_bar_panels', $n( 'add_debug_bar_panel' ) );
+		add_filter( 'debug_bar_statuses', $n( 'add_debug_bar_stati' ) );
+	}
+
 	add_filter( 'ep_formatted_args', $n( 'add_explain_args' ), 10, 2 );
+
+	add_action( 'wp', $n( 'retrieve_raw_document_from_es' ) );
 
 	load_plugin_textdomain( 'debug-bar-elasticpress', false, basename( __DIR__ ) . '/lang' );
 
@@ -151,4 +161,59 @@ function admin_notice_min_ep_version() {
 		</p>
 	</div>
 	<?php
+}
+
+/**
+ * Check if the current page is an indexable singular of an ES document or not.
+ *
+ * @since 3.1.0
+ * @return boolean
+ */
+function is_indexable_singular() {
+	if ( ! is_singular() ) {
+		return false;
+	}
+
+	$id        = get_the_ID();
+	$post_type = get_post_type( $id );
+
+	$post_indexable       = \ElasticPress\Indexables::factory()->get( 'post' );
+	$indexable_post_types = $post_indexable->get_indexable_post_types();
+
+	return in_array( $post_type, $indexable_post_types, true );
+}
+
+/**
+ * Get document from Elasticsearch.
+ *
+ * @since 3.1.0
+ * @return void
+ */
+function retrieve_raw_document_from_es() {
+	if ( empty( $_GET['ep-retrieve-es-document'] ) || empty( $_GET['_wpnonce'] ) || ! wp_verify_nonce( $_GET['_wpnonce'], 'ep-retrieve-es-document' ) ) {
+		return;
+	}
+
+	if ( ! is_indexable_singular() ) {
+		return;
+	}
+
+	\ElasticPress\Indexables::factory()->get( 'post' )->get( get_the_ID() );
+}
+
+/**
+ * Include our QM Output
+ *
+ * @since 3.1.0
+ * @param array $output Array of registered output
+ * @return array
+ */
+function register_qm_output( $output ) {
+	$collector = \QM_Collectors::get( 'elasticpress' );
+
+	if ( $collector ) {
+		$output['elasticpress'] = new QueryMonitorOutput( $collector );
+	}
+
+	return $output;
 }